Fix Config System
The old version cached the config file on initial load and never updated it for the core. It would force Bukkit to reload it, but never update it in the program itself. This caused null pointer issues as the references didn't exist in the object as we should've been using the new config object,